Login page templating
Added template engine to login page as an example for others to follow.

Clearly, the javascript example for setting the menu to highlight the "hof" button is wrong on several fronts: 1) this isn't the hof page; and 2) the user isn't logged in so the {main_nav} template shouldn't even be included. I'm just including for your reference on how to use this elsewhere.