HOTFIX: Most recent event
The SQL query for most recent event was based on date. We clarified with customer on Thursday that there will only be one event active at one time, which would be the most recent event. Thus, it's less a function of the date of the event and more a function of the ID.

We will be removing Date from the event setup.