Database changes
Added create_time and update_time columns in the event table to support cache-busting the champion_photo. @robruwe this affects #29

Changed the date column data type in event from timestamp to date to support the storage of the date selection via the event creation page. @makungaj1 this affects #39.

Changed the photo column in the user table to default to the no-user-image.jpg file for all newly created accounts. @hannonj3 this affects #62.